style: AddPage 디버깅하기 쉽도록 hook 생성하고 코드 분리 - memo 관리, message 만드는 코드를 useMemoManager hook을 만들어서 관리
